+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Changed
   * AZ-166: Use `random_password` instead of `random_string` for passwords generation: **this changes the generated value or need manual state edition**
+  * AZ-182: Allow no \_user suffix
 
 # v2.1.0 - 2019-12-18